+++ frameworks/core/components_ng/base/view_abstract_model.h
@@ -70,9 +70,9 @@ public:
     virtual void SetBackgroundImageSize(const BackgroundImageSize& bgImgSize) = 0;
     virtual void SetBackgroundImagePosition(const BackgroundImagePosition& bgImgPosition) = 0;
     virtual void SetBackgroundBlurStyle(const BlurStyleOption& bgBlurStyle) = 0;
-    virtual void SetSphericalEffect(float radio) {}
+    virtual void SetSphericalEffect(double radio) {}
     virtual void SetPixelStretchEffect(PixStretchEffectOption& option) {}
-    virtual void SetLightUpEffect(float radio) {}
+    virtual void SetLightUpEffect(double radio) {}
 
     virtual void SetPadding(const Dimension& value) = 0;
     virtual void SetPaddings(const std::optional<Dimension>& top, const std::optional<Dimension>& bottom,
